backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
pydantic==2.11.7
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1